Save 2 bytes by reusing the SET_ADDRESS code for SET_CONFIGURATION
`SET_CONFIGURATION` is basically a noop - the only thing that needs to
be done is `process_Host2Device`.  Instead of handling it in two
instructions, reuse the `SET_ADDRESS` code for it, passing the current
UDADDR value to it (so it effectively does nothing too), which can be
done in a single instruction, thus saving 2 bytes.