Fix USB errors due to unconfigured interrupt in endpoint
Although the subset of USB HID protocol that is actually used by the
bootloader uses only the default control endpoint (0), the HID spec
requires the device to have an interrupt in endpoint, and the host can
poll that endpoint even when the HID report descriptor does not actually
declare any input reports.  Polling an unconfigured endpoint causes USB
errors, which may prevent the bootloader from functioning properly.

Add the code to configure Endpoint1 as Interrupt IN, matching the USB
descriptors; this is enough to make the USB controller generate NAK
replies for that endpoint correctly, and the rest of bootloader code may
continue using just Endpoint0 as before.

The binary size increases by 12 bytes.