Save 2 bytes by refactoring the GET_DESCRIPTOR code
Rewrite the part of the GET_DESCRIPTOR handling code that loads the
address and length of the requested descriptor to use less jumps; the
resulting code consumes 2 bytes less, even though it is actually more
correct (no longer replies with some descriptor to requests for an
unknown descriptor type).  The new code also avoids hardcoding the high
address byte, and no longer depends on the fact that all descriptors
have the same high address byte (but it depends on the fact that all
offsets between adjacent descriptors can fit into the `adiw` constant
argument (0...63)).