Save 2 bytes in the flash write loop
Instead of saving the initial address value in a separate register pair,
and then reloading it after the data write loop, keep the address in the
Z register and subtract 128 from it after the data write loop - this
should give the same result, because the loop counter is explicitly
initialized by the bootloader code.  However, because the `sbiw`
instruction supports only the 0...63 range for its constant argument,
another optimization trick is also used - instead of incrementing the
whole 16-bit address, only the low byte is incremented; this should give
the same result, because the block start address must be aligned to the
flash page size (128 bytes).