Add support for starting the bootloader on application request
The startup code of nanoBoot checked only the EXTRF bit in MCUSR when
determining whether the bootloader or the application should start.
However, in many cases the application may need a way to start the
bootloader even when the actual external reset did not happen (e.g.,
after some "upgrade firmware" action in the user interface provided by
the application, or even after receiving some special request over USB,
although that might have some security implications).  Because the reset
source bits in MCUSR cannot be changed from 0 to 1 by software, there
was no way around the reset source check in the nanoBoot code (apart
from jumping to the code after the check directly, which is really
fragile and should be avoided).

Several existing bootloaders use the `MCUSR == 0` check in some form to
detect the application request to enter the bootloader.  Add the same
kind of code to nanoBoot (but instead of a simple comparison, use a mask
which covers all MCUSR bits defined in the appropriate datasheet, so
that the code won't give wrong results if some reserved MCUSR bits do
not read as 0 for some reason).

The binary size increases by 6 bytes.