Save 4 bytes in the USB detach code
The original LUFA code performed `UDCON |= (1 << DETACH);` to detach the
USB device; however, the other bits of `UDCON` are known to be 0 at this
time, therefore a simple write of a constant value could be performed
here.  In addition, the value of `(1 << DETACH)` is 1 on all AVR chips
that could be potentially supported by this code, therefore even the
instruction to load the constant value into a register could be omitted.
Doing these changes removes 2 instructions, saving 4 bytes.