Save 2 bytes in the HOST_TO_DEVICE parsing code
The instruction which applied the mask of `(CONTROL_REQTYPE_TYPE |
CONTROL_REQTYPE_RECIPIENT)` to bmRequestType was not actually needed,
because the value of this mask is 0x7F, and the only bit which is not
covered by the mask (0x80) is already known to be 0, therefore the
masking did not actually change anything.  Removing this instruction
saves 2 bytes.

Signed-off-by: Sergey Vlasov <sigprof@gmail.com>

Since following commits are skipped previously when cherry-picking.
  97c8d96 ("TEMPORARY: EEPROM code which does not fit", 2021-08-25)
  e48801a ("WIP: Optimize some more code to make the EEPROM support fit", 2021-08-25)
  78b804d ("WIP: Redo the EEPROM write implementation", 2021-08-29)
  e7e94f3 ("Save 2 bytes by reusing the TXINI clearing mask", 2021-08-29)

Adjusted to keep branch to thunk code.

Signed-off-by: Osamu Aoki <osamu@debian.org>