Save 2 bytes in the HOST_TO_DEVICE parsing code
The instruction which applied the mask of `(CONTROL_REQTYPE_TYPE |
CONTROL_REQTYPE_RECIPIENT)` to bmRequestType was not actually needed,
because the value of this mask is 0x7F, and the only bit which is not
covered by the mask (0x80) is already known to be 0, therefore the
masking did not actually change anything.  Removing this instruction
saves 2 bytes.