Handle GET_DESCRIPTOR with wLength > 255 correctly
The host can send a GET_DESCRIPTOR request with wLength > 255; in this
case comparing just the low byte of wLength with the descriptor length
might not give the correct result.  Check the high byte of wLength too,
so that those requests (even if they are unusual) would be handled
correctly.

This change adds one extra instruction (2 bytes).