Save 4 bytes by reusing process_Host2Device for SET_HID_REPORT
The logic at the end of SET_HID_REPORT handling was almost the same,
apart from the final wait for TXINI, which was not done in the
SET_HID_REPORT case.  If the wait is always performed, the same code
could be used for both SET_HID_REPORT and no-data control requests,
saving 4 bytes.

A potential problem here is that a lost ACK packet from host which was
sent as a response to the DATA1 packet from the device at the end of the
status stage may now result in a lockup; however, the same problem was
already affecting the no-data control requests (SET_ADDRESS and
SET_CONFIGURATION).