Save 6 bytes by refactoring UNHANDLED_SETUP_REQUEST usage
Remove the code which tried to check whether the SETUP packet has been
handled - now UNHANDLED_SETUP_REQUEST can be used only if the SETUP
packed needs to be acknowledged and replied with STALL, and in other
cases the code should just do `reti` directly.  In addition,
UNHANDLED_SETUP_REQUEST is placed in the middle of the code, so that it
would be reachable by conditional branches directly.

Signed-off-by: Sergey Vlasov <sigprof@gmail.com>

This is based on
6bc21dd ("Save 6 bytes by refactoring UNHANDLED_SETUP_REQUEST usage", 2021-09-04)

UNHANDLED_SETUP_REQUEST is moved up so we don't need to branch through
thunk like UNHANDLED_SETUP_REQUEST_1.

So drop UNHANDLED_SETUP_REQUEST_1 usage.

Signed-off-by: Osamu Aoki <osamu@debian.org>